Gif Squad

+++ src/main/java/com/banco/bluebank/service/AgenciaService.java
@@ -2,6 +2,7 @@ package com.banco.bluebank.service;
 
 import com.banco.bluebank.exceptionhandler.exceptions.AgenciaNaoEncontradaException;
 import com.banco.bluebank.exceptionhandler.exceptions.EntidadeEmUsoException;
+import com.banco.bluebank.exceptionhandler.exceptions.RecursoComBloqueioException;
 import com.banco.bluebank.model.Agencia;
 import com.banco.bluebank.repository.AgenciaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,11 +17,15 @@ public class AgenciaService {
 
     private static final String MSG_AGENCIA_EM_USO
             = "Agência de id %d não pode ser removida, pois está em uso";
+    public static final long AGENCIA_GLOBAL = 1L;
 
     @Autowired
     private AgenciaRepository agenciaRepository;
 
     public Agencia salvar(Agencia agencia){
+        if(agencia.getId()!=null && agencia.getId()== AGENCIA_GLOBAL)
+            throw new RecursoComBloqueioException("Recurso de agência bloqueado para esta operacão por motivo de manter a integridade da regra de negócios");
+
         return agenciaRepository.save(agencia);
     }
 
@@ -34,6 +39,9 @@ public class AgenciaService {
     }
 
     public void excluir(Long agenciaId) {
+        if(agenciaId == AGENCIA_GLOBAL)
+            throw new RecursoComBloqueioException("Recurso de agência bloqueado para esta operacão por motivo de manter a integridade da regra de negócios");
+
         try {
 
             agenciaRepository.deleteById(agenciaId);
